Possible fix

+++ Program4.java
@@ -79,7 +79,7 @@ public class Program4 extends Tokenizer
 
 	public void printComparisons()
 	{
-		int average = (this.sum / this.keys.length);
+		long average = (this.sum / this.keys.length);
 
 		System.out.print("\n");
 		System.out.println("Size of set: "+this.words.count());
